Split Fixed Operating Cost Plugin into two: - Labor Operating Cost Plugin, and Other Fixed Operating Cost Plugin - Update the test suite (plugin tests) as well as the Default workflow accordingly
