+++ src/pyH2A/Config/Defaults.md
@@ -9,10 +9,11 @@ initial_equity_depreciable_capital | function | 4 | core function to process dep
 non_depreciable_capital_costs | function | 5 | core function to process non-depreciable capital costs
 Replacement_Plugin | plugin | 6 | Calculation of replacement costs
 replacement_costs | function | 7 | core function to process replacement costs
-Fixed_Operating_Cost_Plugin | plugin | 8 |Calculation of fixed operating costs
-fixed_operating_costs | function | 9 | core function to process fixed operating costs
-Variable_Operating_Cost_Plugin | plugin | 10 | Calculation of variable operating costs, including utilities
-variable_operating_costs | function | 11 | core function to process variable operating costs
+Labor_Operating_Cost_Plugin | plugin | 8 |Calculation of labor operating costs
+Other_Fixed_Operating_Cost_Plugin | plugin | 9 |Calculation of other fixed operating costs
+fixed_operating_costs | function | 10 | core function to process fixed operating costs
+Variable_Operating_Cost_Plugin | plugin | 11 | Calculation of variable operating costs, including utilities
+variable_operating_costs | function | 12 | core function to process variable operating costs
 
 # Financial Input Values
